pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "TheOriginalsTest"
version = "0.1.0"
description = "The Originals Test"
authors = ["Tsezar"]

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = ">=2.0.19,<3.0"
alembic = ">=1.11.1,<2.0"
python-decouple = ">=3.8,<4.0"
python-jose = ">=3.3.0,<4.0"
passlib = ">=1.7.4,<2.0"
uvicorn = ">=0.24.0,<0.25"
asyncpg = ">=0.29.0,<0.30"
fastapi = ">=0.109.0,<0.110"
sqlalchemy-utils = ">=0.41.1,<0.42"
pyjwt = "==2.8.0"
email-validator = ">=2.0"
celery = ">=5.0.5,<6.0"
redis = ">=4.0,<5.0" 
aiosmtplib = ">=1.1,<2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
